Remove hushh profile search and shadow APIs
