--- src/services/profileSearch/index.ts
@@ -1,6 +1,0 @@
-// Profile Search Service - Entry Point
-// Export all profile search functionality
-
-export { searchProfile, mapToOnboardingFields } from './profileSearchService';
-export type { SearchParams, ProfileSearchResult } from './profileSearchService';
-export * from './types';
