--- supabase/functions/hushh-profile-search/deno.json
@@ -1,5 +1,0 @@
-{
-  "imports": {
-    "blueimp-md5": "npm:blueimp-md5@^2.19.0"
-  }
-}
