--- supabase/functions/shadow-investigator/deno.json
@@ -1,6 +1,0 @@
-{
-  "compilerOptions": {
-    "lib": ["deno.ns", "deno.unstable"],
-    "strict": true
-  }
-}
